requirements.txt
flask==3.0.3
flask-socketio==5.4.1
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
dnstwist==20250130
beautifulsoup4==4.12.3
requests==2.32.3
Levenshtein==0.26.1
python-whois==0.9.6
flask-login==0.6.3